requirements.txt
python-telegram-bot
pandas
openpyxl
python-dotenv
requests
